+++ src/dd/meta/modules.js
@@ -1,3 +1,0 @@
-config({
-    'dd': {alias: ['dd/base', 'dd/droppable']}
-});