+++ src/separator/src/separator/render.js
@@ -1,0 +1,15 @@
+/**
+ * separator render def
+ * @author yiminghe@gmail.com
+ */
+KISSY.add("separator/render", function (S, Component) {
+
+    return Component.Render.extend({
+        initializer: function () {
+            this.get('elAttrs')['role'] = 'separator';
+        }
+    });
+
+}, {
+    requires: ['component/base']
+});